+++ src/IRremote.h
@@ -63,17 +63,16 @@
  * Each protocol you include costs memory and, during decode, costs time
  * Disable (deactivate the line by adding a trailing comment "//") all the protocols you do not need/want!
  */
-#if (!(defined(DECODE_DENON) || defined(DECODE_SHARP) || defined(DECODE_JVC) || defined(DECODE_KASEIKYO) \
+#if (!(defined(DECODE_DENON) || defined(DECODE_JVC) || defined(DECODE_KASEIKYO) \
 || defined(DECODE_PANASONIC) || defined(DECODE_LG) || defined(DECODE_NEC) || defined(DECODE_SAMSUNG) \
 || defined(DECODE_SONY) || defined(DECODE_RC5) || defined(DECODE_RC6) || defined(DECODE_HASH) \
 || defined(DECODE_BOSEWAVE) || defined(DECODE_LEGO_PF) || defined(DECODE_MAGIQUEST) || defined(DECODE_WHYNTER)))
-#define DECODE_DENON
-#define DECODE_SHARP        // the same as DECODE_DENON
+#define DECODE_DENON        // Includes Sharp
 #define DECODE_JVC
 #define DECODE_KASEIKYO
 #define DECODE_PANASONIC    // the same as DECODE_KASEIKYO
 #define DECODE_LG
-#define DECODE_NEC
+#define DECODE_NEC          // Includes Apple and Onkyo
 #define DECODE_SAMSUNG
 #define DECODE_SONY
 #define DECODE_RC5